backend/requirements.txt
# ─── Web Framework ──────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# ─── Supabase ────────────────────────────────────────────────────────────────
supabase>=2.4.0

# ─── AI / ML ────────────────────────────────────────────────────────────────
openai-whisper
torch
torchaudio
librosa>=0.10.0
scipy>=1.12.0
google-generativeai>=0.7.0

# ─── Video Processing ───────────────────────────────────────────────────────
mediapipe>=0.10.0
opencv-python>=4.9.0
moviepy>=1.0.3

# ─── Utilities ──────────────────────────────────────────────────────────────
numpy>=1.26.0
